✨ feat: select network page
