+++ apps/wallet/src/pages/main/index.tsx
@@ -4,14 +4,15 @@ import SendButton from "../../components/SendButton";
 import ReceiveButton from "../../components/ReceiveButton";
 import TokenList from "../../components/TokenList";
 import { useTokenListQuery } from "../../apis/react-query/token";
-import ChainSelect from "../../components/ChainSelect";
 import hibitIdSession from "../../stores/session";
 import { useQuery } from "@tanstack/react-query";
 import { ChainAssetType } from "../../utils/basicTypes";
 import { formatAddress } from "../../utils/formatter";
 import CopyButton from "../../components/CopyButton";
 import SvgSettings from '../../assets/setting.svg?react';
-import { useNavigate } from "react-router-dom";
+import { Link, useNavigate } from "react-router-dom";
+import ChainIcon from "../../components/ChainIcon";
+import SvgCaret from '../../assets/caret-down.svg?react'
 
 const WalletMainPage: FC = observer(() => {
   const navigate = useNavigate()
@@ -31,12 +32,13 @@ const WalletMainPage: FC = observer(() => {
   return (
     <div className="h-full px-6 relative flex flex-col gap-6 overflow-auto">
       <div className="absolute top-0 left-6">
-        <ChainSelect
-          value={hibitIdSession.chainInfo}
-          onChange={(chain) => {
-            hibitIdSession.switchChain(chain)
-          }}
-        />
+        <Link
+          to="/network-select"
+          className="btn btn-xs px-0 pr-1 gap-1 rounded-full bg-neutral"
+        >
+          <ChainIcon chainInfo={hibitIdSession.chainInfo} size="sm" onlyIcon />
+          <SvgCaret />
+        </Link>
       </div>
       <button className="btn btn-ghost btn-square btn-sm hover:bg-transparent absolute -top-1 right-6" onClick={() => {
         navigate('/settings')
